211122_1

+++ sinatra_srv.rb
@@ -260,10 +260,14 @@ get /\/work\/convert\/(.*)\/(.*)\/(.*)\/(.*)\/(.*)\/(.*)\/(.*)\// do
 
     msg = e.message
 
-
-    pid_new = spawn("ruby C:/www/root/sinatra_1c_77/sinatra_srv.rb -e production")
-
-     Process.kill("TERM", main_pid)
+    if pid == ""
+      p "***********************************************************************************"
+      p "Try to create new application..."
+      pid_new = spawn("ruby C:/www/root/sinatra_1c_77/sinatra_srv.rb -e production")
+      p "NEW APP PID - #{pid_new}"
+      p "kill old application..."
+      Process.kill("TERM", main_pid)
+    end
 
     return msg.to_s
